Create Small Lottery Game (in Java)
Design a small lottery game. The system should produce a two-digit number. You should also ask the user to guess this number. If the user guesses the number exactly correctly, he wins $10000, 
If the user knows the digits of the number, i.e. if he wrote 56 instead of 65, he wins $5000, If the user knows only one digit of the lucky number, he wins $1000.